+++ test/helpers.js
@@ -108,6 +108,16 @@ const createERC721DepositProposalData = (
   ); // metaData                      (?? bytes)
 };
 
+const createBtcDepositData = (
+  transferAmount,
+  btcRecipientAddress
+) => {
+  return Ethers.utils.solidityPack(
+    ["uint256", "uint256", "string"],
+    [transferAmount, btcRecipientAddress.length, btcRecipientAddress]
+  )
+}
+
 const advanceBlock = () => {
   const provider = new Ethers.providers.JsonRpcProvider();
   const time = Math.floor(Date.now() / 1000);
@@ -382,7 +392,7 @@ const reverts = async function(promise, expectedErrorMessage) {
     await promise;
   } catch (error) {
     if (expectedErrorMessage) {
-      const message = error.reason || error.hijackedStack.split('revert ')[1].split('\n')[0];
+      const message = error.reason || error.hijackedStack.split("revert ")[1].split("\n")[0];
       assert.equal(message, expectedErrorMessage);
     }
     return true;
@@ -416,6 +426,7 @@ module.exports = {
   createERC1155DepositProposalData,
   createERC1155WithdrawData,
   createGmpDepositData,
+  createBtcDepositData,
   constructGenericHandlerSetResourceData,
   createERC721DepositProposalData,
   createResourceID,